requirements.txt
Django==1.10.2
djangorestframework==3.5.1
requests==2.11.1
